Seismic Facies classification using Deep Learning techniques.
Earthquakes around the world have been a cause of major destruction and loss of life and property. Seismic reflection data are a key source of information in numerous fields of geoscience. However, the often subjective and nonunique interpretation of seismic reflection data has led to longstanding debates based on contrasting geologic interpretations of the same or similar data sets Moreover, seismic interpretations require significant amounts of time, experience, and expertise from interpreters.
The earthquake stations continuously collect data even when there is no event. From this data, we need to distinguish earthquake and non-earthquake. An early detection and prediction system using classification models can prove to be very useful for disaster management teams. 
Aim is to build a machine-learning model that can predict the facies classifications for each pixel.